requirements.txt
absl-py==2.4.0
altgraph==0.17.5
astunparse==1.6.3
backports.tarfile==1.2.0
blinker==1.9.0
build==1.4.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
docutils==0.22.4
Flask==3.1.3
flask-cors==6.0.2
Flask-JWT-Extended==4.7.1
Flask-Mail==0.10.0
flatbuffers==25.12.19
fpdf==1.7.2
gast==0.4.0
google-auth==2.49.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.78.0
h5py==3.14.0
id==1.6.1
idna==3.11
importlib_metadata==8.7.1
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.1.1
jaraco.functools==4.4.0
jax==0.4.30
jaxlib==0.4.30
Jinja2==3.1.6
keras==3.13.2
keyring==25.7.0
libclang==18.1.1
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.4
more-itertools==10.8.0
mysql-connector-python==9.6.0
namex==0.1.0
nh3==0.3.3
numpy==2.4.3
oauthlib==3.3.1
opt_einsum==3.4.0
optree==0.19.0
packaging==26.0
pefile==2024.8.26
pillow==12.1.1
protobuf==7.34.0
psycopg2-binary==2.9.11
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
Pygments==2.19.2
pyinstaller==6.19.0
pyinstaller-hooks-contrib==2026.3
PyJWT==2.11.0
pyproject_hooks==1.2.0
python-dotenv==1.2.1
pywin32-ctypes==0.2.3
readme_renderer==44.0
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==14.3.3
scipy==1.15.3
six==1.17.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.21.0
tensorflow-estimator==2.15.0
tensorflow-intel==0.0.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.3.0
twine==6.2.0
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.6
wrapt==1.14.2
zipp==3.23.0
ai-edge-litert==2.1.3
